Move landing_page module where it belongs
