+++ usmqe/web/tendrl/landing_page/pages.py
@@ -11,7 +11,7 @@ Author: ltrilety
 import copy
 
 from usmqe.web.tendrl.auxiliary.pages import UpperMenu
-import usmqe.web.tendrl.mainpage.landing_page.models as m_landing_page
+import usmqe.web.tendrl.landing_page.models as m_landing_page
 from usmqe.web.tendrl.mainpage.navpage.pages import NavMenuBars
 from usmqe.web.tendrl.mainpage.clusters.cluster_list.pages\
     import ClustersWorkBase
